[기능 수정] 리팩토링
ViewController에 model과 view를 한꺼번에 구현하였으나, 이를 model 클래스 따로, view 클래스 따로 생성하여 분리했습니다.
