+++ PhotosApp/PhotosApp/PhotosView.swift
@@ -1,0 +1,14 @@
+//
+//  PhotosView.swift
+//  PhotosApp
+//
+//  Created by 신한섭 on 2020/03/16.
+//  Copyright © 2020 신한섭. All rights reserved.
+//
+
+import UIKit
+
+class PhotosView: UIView {
+    
+    @IBOutlet weak var collectionView: UICollectionView!
+}
