+++ PhotosApp/PhotosApp/ViewController.swift
@@ -9,39 +9,17 @@
 import UIKit
 
 class ViewController: UIViewController {
-
-    @IBOutlet weak var collectionView: UICollectionView!
-        
+    
+    private let dataManager = DataManager()
+    private var photosView: PhotosView {
+        view as! PhotosView
+    }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
-        collectionView.delegate = self
-        collectionView.dataSource = self
-    }
-    
-    private func generateRandomColor() -> UIColor {
-        let range: Range<CGFloat> = (1..<255)
-        let red = CGFloat.random(in: range) / 255
-        let green = CGFloat.random(in: range) / 255
-        let blue = CGFloat.random(in: range) / 255
-        let alpha = CGFloat.random(in: 0.5..<1)
-        return UIColor(red: red, green: green, blue: blue, alpha: alpha)
-    }
-    
-}
-
-extension ViewController: UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 4
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "collectionViewCell", for: indexPath)
-        let color = generateRandomColor()
-        cell.backgroundColor = color
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: 80, height: 80)
+        photosView.collectionView.dataSource = dataManager
+        photosView.collectionView.delegate = dataManager
     }
 }
+    
+
